// frontend/src/api/user_api.ts
export interface verificationQuestionType {
    problem1?: string
    problem2?: string
    problem3?: string
    answer1?: string
    answer2?: string
    answer3?: string
}

// 用户配置信息
export interface userConfType {
    "userID": number
    "nickname": string
    "abstract": string
    "avatar": string
    "recallMessage"?: string // 撤回消息的提示词
    "friendOnline": boolean // 好友上线提醒
    "sound": boolean // 声音
    "secureLink": boolean // 安全连接
    "savePwd": boolean // 报存密码
    "searchUser": number
    "verification": number
    "verificationQuestion"?: verificationQuestionType
}