// frontend/src/api/group_api.ts
export interface groupInfoType {
    "groupId": number
    "title": string
    "abstract": string
    "memberCount": number
    "memberOnlineCount": number
    "avatar": string
    "creator": userBaseInfoType
    "adminList": userBaseInfoType[]
    "role": number
    "isProhibition": boolean //是否全员禁言
    "prohibitionTime"?: number
    "isSearch": boolean
    "isInvite": boolean
    "isTemporarySession": boolean
}
// ...
export interface groupUpdateRequest {
    "id": number
    "avatar"?: string
    "isProhibition"?: boolean
    "title"?: string
    "abstract"?: string
    isSearch?: boolean // 是否被搜索
    verification?: number // 群验证 0 不允许任何人添加  1 允许任何人添加  2 需要验证消息 3 需要回答问题  4  需要正确回答问题
    isInvite?: boolean // 是否可邀请好友
    isTemporarySession?: boolean // 是否开启临时会话
    verificationQuestion?: verificationQuestionType
}
// ...
export interface groupHistoryType {
    groupID: number
    "userID": number
    "userNickname": string
    "userAvatar": string
    "msg": msgType
    "id": number
    "msgType": number
    "createdAt": string
    "isMe": boolean
    "memberNickname": string
    msgPreview: string
    showDate: boolean
}